+++ react/__mocks__/vtex.styleguide.js
@@ -2,19 +2,26 @@ import React, { Component } from 'react'
 
 export class Dropdown extends Component {
   render() {
-    return (<div>Dropdown</div>)
+    return <div>Dropdown</div>
   }
 }
 
 export class Spinner extends Component {
   render() {
-    return (<div>Spinner</div>)
+    return <div>Spinner</div>
   }
 }
 
-
 export class Checkbox extends Component {
   render() {
-    return (<div>Checkbox</div>)
+    return <div>Checkbox</div>
   }
+}
+
+export function IconCaretLeft() {
+  return <div>IconCaretLeft</div>
+}
+
+export function IconCaretRight() {
+  return <div>IconCaretRight</div>
 }