+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `SearchFooter` and `SearchResultInfiniteScroll` tests.
+
 ### Changed
 - Link redirect working based on the `pagesPath` prop
+
+## [0.4.0] - 2018-6-20
+### Added
+- Add the Infinite Scroll feature to the page.
+- Add a spinner to be shown when fetching more products.
 
 ### Fixed
 - Fix `SearchFilter` title internationalization.